+++ src/main/kotlin/christmas/domain/menu/Menu.kt
@@ -1,4 +1,4 @@
-package christmas.domain.order
+package christmas.domain.menu
 
 import christmas.domain.Money
