+++ src/main/kotlin/christmas/domain/menu/MenuCount.kt
@@ -1,4 +1,4 @@
-package christmas.domain.order
+package christmas.domain.menu
 
 import christmas.exception.OrdersException
