+++ src/test/kotlin/christmas/domain/menu/OrdersTest.kt
@@ -1,4 +1,4 @@
-package christmas.domain.order
+package christmas.domain.menu
 
 import christmas.domain.Money
 import org.assertj.core.api.Assertions
